ref: Optimize file fetching code (#96)
While looking around in here I noticed we were doing a whole extra fetch
to get the metadata for the file when all of the info we need was
already present on the page. This PR updates the code to remove that
fetch. Additionally, adds support for file coverage when browsing at a
specific commit, this was previously not supported.