feat: Build support for new github pr experience (#152)
Adds support for the new github PR experience. This could have been a
huge refactor, but I decided to do it more surgically so as to not let
this blow out of proportion. Tested the whole matrix of new (new, old),
(unified, split), (chrome, ff). Will test again on main before release.

Closes https://github.com/codecov/codecov-browser-extension/issues/151